[MODINVOSTO-73] - Refactor packages and tests updated
