+++ src/main/java/org/folio/rest/utils/RestConstants.java
@@ -1,4 +1,4 @@
-package org.folio.rest.util;
+package org.folio.rest.utils;
 
 public final class RestConstants {
   public static final String OKAPI_URL = "x-okapi-url";
