+++ src/main/java/org/folio/service/BatchVoucherExportsService.java
@@ -3,7 +3,7 @@ package org.folio.service;
 import static com.google.common.collect.ImmutableMap.of;
 import static javax.ws.rs.core.Response.Status.NOT_FOUND;
 import static org.folio.rest.impl.BatchVoucherExportsImpl.BATCH_VOUCHER_EXPORTS_TABLE;
-import static org.folio.rest.util.ResponseUtils.handleNoContentResponse;
+import static org.folio.rest.utils.ResponseUtils.handleNoContentResponse;
 
 import java.util.Map;
